Refactored components to have index files
